finalize recommendation system
