+++ docs/mini-projects/recommendation-system/intro-to-rec-systems.md
@@ -1,0 +1,42 @@
+---
+jupytext:
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.14.6
+kernelspec:
+  display_name: jupyblog
+  language: python
+  name: python3
+---
+
+# An introduction to Recommendation Systems
+
+Recommendation systems have become an integral part of our digital lives. From suggesting movies on Netflix to recommending products on Amazon, these systems play a pivotal role in enhancing user experience and driving business metrics. In this blog, we'll delve deep into the world of recommendation systems, focusing on content-based recommenders.
+
+## What are Recommendation Systems?
+
+Recommendation systems are algorithms designed to suggest relevant items to users. These items could be anything from movies, songs, and books to products, news articles, or even search queries. The primary goal is to provide personalized recommendations that enhance the user's experience.
+
+## Why are they Important?
+
+In today's digital age, the sheer volume of choices can be overwhelming. Recommendation systems help users navigate through this vast sea of information, providing them with tailored suggestions that align with their preferences and behaviors. This not only enhances user satisfaction but also boosts business metrics like sales, retention, and engagement.
+
+## Types of Recommendation Systems
+
+Recommendation systems can be broadly classified into two categories:
+
+* **Collaborative Filtering**: This method uses the past behavior of users and items to generate recommendations. For instance, if a user has previously purchased a book, the system might recommend similar books based on the purchase history of other users who have bought the same book.
+* **Content-Based Filtering**: This method uses the attributes of items to recommend other items with similar attributes. For instance, if a user has previously purchased a book, the system might recommend similar books based on the genre, author, or publisher of the book.
+* **Matrix Factorization**: Matrix factorization techniques, like Singular Value Decomposition (SVD), decompose the user-item interaction matrix into multiple matrices representing latent factors. It's especially popular for its effectiveness in collaborative filtering and its ability to handle sparse data.
+* **Hybrid Methods**: hybrid methods combine the strengths of both collaborative and content-based filtering. Hybrid models can be implemented in several ways:
+
+1. Separate models for each approach that are combined at the end.
+2. Incorporating collaborative and content-based methods into a single model.
+3. Unifying the models into a single model.
+
+* **Deep Learning**: With the rise of neural networks and deep learning, these techniques are increasingly being applied to recommendation systems. Autoencoders or Recurrent Neural Networks (RNNs) can be used to predict the next item in a sequence, making them useful for recommending items like the next song in a playlist or the next video in a series.
+* **Association Rule Mining**: Used mainly in market basket analysis, this method identifies associations between items. A classic example is the association between diapers and beer in a supermarket setting. Tools like Apriori or Eclat algorithms are used to extract these associations.
+* **Knowledge-Based Recommendations**: In situations where user-item interactions are sparse and it's challenging to compute reliable recommendations, knowledge-based techniques come in handy. They provide personalized recommendations by leveraging explicit knowledge about users and items. They often involve asking users to give more input or provide feedback.
+* **Session-Based Recommendations**: In many scenarios, especially in e-commerce, the user might not have a past history, or their current intent might be different from their past behavior. Session-based recommenders focus on short-term behavior and use techniques like RNNs to predict the next item a user might be interested in during a session.
